wellness-backend/requirements.txt
fastapi
uvicorn[standard]
beanie
motor
pydantic-settings
email-validator
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
resend
openai
boto3
python-multipart
pytest
httpx
pytest-asyncio
streamlit
